Add files via upload
5V Regulator



Is to keep a constant output voltage even if the input voltage changes, why is it important because electronic components can only handle a certain voltage, so we need a way to maintain a constant output voltage even when the input voltage changes, let's say a constant 5 volt dc supply and simple, we want to provide enough current to charge a cheap phone, we want to be able to connect it to multiple voltage sources like 9 volt or maybe 12 volt batteries, to get it we need to use an integrated circuit component, there are many options, all of which can operate at different voltages, but after a bit of research I found this, lm7805 this is a fixed 5 volt dc output and current up to 1.5 amps this component can be connected to any dc supply voltage between 7 and 35 volts so perfect for our needs it is 3 pin pin 1 unregulated voltage input pin 2 is ground pin and pin 3 , is the regulated 5 volt output, the manufacturer recommends a capacitor at the input and notes at the output that the input capacitor is needed if the regulator is far from the power supply filter, we will use some long wires. Connect the battery so we will use the recommended 0.22 microfarad capacitor, this is an electrolytic type capacitor, we can use a slightly larger capacity version but we don't want to use a smaller capacitor, we will add another capacitor parallel to the input side that will help smooth the capacitor supply interruptions this is a bypass capacitor this regulator is placed very close to the input pin this will be a small ceramic capacitor typically 0.1 microfarad the purpose this capacitor is to filter out noise and high frequency distortions from the power supply because we may not always get a perfectly flat dc supply also another 0 to the output side We will add a 1 microfarad bypass capacitor and a 10 microfarad electrolytic capacitor. just a typical value used for this purpose, if we want we can use a slightly higher capacity version, but this will work fine, these will help us to get a clean output in our connected circuit, we will also add a protection diode to the input on the hand this will protect the circuit if we connect the power supply incorrectly it helps to show how it works if i connect this incandescent lamp to a power source i can reverse the wires that will burn and if i put a diode on it it will also light up red wire and connect it to positive it will light up again but now when i reverse the leads the diode blocks the current and the lamp stays off so i do this we can use it to protect the circuit, connected a 9 volt battery to the source and the multimeter shows 5 volts in the socket, if I reverse the battery, we see zero volts, the multimeter, that is, the diode protects our circuit, I am happy with it, so I put a small load on it. Circuit diagram, gerber file and boom list are attached.